requirements.txt
Pillow==7.1.2
opencv-python==4.2.0.34
matplotlib==3.2.1
scipy==1.4.1
jupyterlab==2.1.2